# Avoid panics in all heap functions