Allows the compiler to infer that the core heap methods can't panic due to bounds checks. (`push_pop_{max, min}()` and `replace_{max, min}()` still need to be replaced with versions using `peek_{min, max}_mut()` to avoid bounds checks. I implemented this in my previous PR: #8.) This significantly improves performance, as the compiler avoids branch instructions, outputs less code, and is able to inline the functions completely.

Before:
```
test pop_max_seq ... bench:      69,594 ns/iter (+/- 9,316)
test pop_min_seq ... bench:      86,875 ns/iter (+/- 2,483)
test push_seq    ... bench:      22,185 ns/iter (+/- 1,641)
```
After:
```
test pop_max_seq ... bench:      64,055 ns/iter (+/- 5,473)
test pop_min_seq ... bench:      86,439 ns/iter (+/- 2,132)
test push_seq    ... bench:      15,420 ns/iter (+/- 2,914)
```