@calebsander 2021-09-13
Yes, they should be able to work together. They could be merged into `master` in either order; once one of the PRs is merged, I will fix the merge conflicts in the other, specifically:
- The implementations of `push_pop_{max, min}()` and `replace_{max, min}()` can be taken entirely from #8, since it implements them in terms of other heap methods
- The `Drop` implementations for `Peek{Max, Min}Mut` can be merged manually; this PR only changes the code for doing the trickle and #8 changes the code for deciding *when* to do the trickle